Feature/parallel loading (#1)
* Load files in parallel

This results in a huge speedup on my laptop, from 2min40 to 1min40 when
loading all the files.

* Added parallel loading for R

`mclapply` is used, which only works on Linux/Mac, rather than
`parLapply`, which is multiplatform. This is because the latter approach
did not result in any speed ups during development over a single core
approach.