Add table to store corrections to ref measurements
The ref_raw table should now contain all raw reference measurements, and
if any corrections are to be applied (such as setting a calibration
value to NULL) then this should be done by adding a corresponding entry
in the ref_corrections table. This is more transparent than the previous
method of writing corrections directly to the database and making it
harder to follow where these had been done.