Change update scripts to use ref_corrections
Rather than directly modifying suspicious values in the ref_raw tables,
they are instead added to the ref_corrections table. Currently this
table contains a full copy of all Manchester reference data from 2021
onwards as the CO2 data for this period is unreliable, and a new row
needs to be the ref_corrections table even if only 1 field needs
correcting.