Feature/reference corrections separate table (#8)
* Add table to store corrections to ref measurements

The ref_raw table should now contain all raw reference measurements, and
if any corrections are to be applied (such as setting a calibration
value to NULL) then this should be done by adding a corresponding entry
in the ref_corrections table. This is more transparent than the previous
method of writing corrections directly to the database and making it
harder to follow where these had been done.

* Change update scripts to use ref_corrections

Rather than directly modifying suspicious values in the ref_raw tables,
they are instead added to the ref_corrections table. Currently this
table contains a full copy of all Manchester reference data from 2021
onwards as the CO2 data for this period is unreliable, and a new row
needs to be the ref_corrections table even if only 1 field needs
correcting.

* Add O3, NO2, NO ref York data to DB

The NO2 and NO series contain significant calibration artefacts and so
have been removed from the user-facing `ref` view using the
`ref_corrections` table.