LCS view show most recent contemporary data
I.e. for a given timepoint, if there are the choice of multiple
calibration versions comprising the 'live' version that was available at
the time, in addition to any calibrations that were retrospectively
applied, the LCS view displays the 'live' contemporary value.
Previously it returned the latest calibration that was retrospectively
applied.

This change means the LCS view represents the contemporary measurement,
which is more in line with what a user would expect. To make this
explicit, the calibrationversion name is returned as a field